requirements.txt
selenium==4.11.2
selenium-stealth==1.0.6
webdriver-manager==4.0.1
requests==2.31.0
pyyaml==6.0.1
python-xlib==0.33
python-uinput==0.11.2
